Add files via upload
First the word document is parsed.
The document is cleaned by removing the spaces and it is converted into tokens of size 512 and the Number of token  the token data are printed.
These token are convert into embeddings and these embeddings are stored in vector database. here we have used Fiass database.
The chunks which we have got as a output are parsed as a into to the chunk array. 
So, when we ask the question it will return the Top 3 matches for the given question.